% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_dados_divisao.R
\name{comercioasia_dados_divisao}
\alias{comercioasia_dados_divisao}
\title{Dados de fluxo de comércio bilateral entre Brasil e um determinado país}
\usage{
comercioasia_dados_divisao(periodo)
}
\arguments{
\item{periodo}{"anual" ou "mensal"}

\item{pais}{um país}
}
\value{
tibble com dados de comércio bilateral de 2010 em diante.
}
\description{
Dados de fluxo de comércio bilateral entre Brasil e um determinado país
}
