% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_dados_omc.R
\name{comercioasia_dados_omc}
\alias{comercioasia_dados_omc}
\title{Dados extraidos da base_omc}
\usage{
comercioasia_dados_omc(pais)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual"}
}
\description{
Dados extraidos da base_omc
}
