% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_get_ulimoano.R
\name{comercioasia_get_ulimoano}
\alias{comercioasia_get_ulimoano}
\title{Último ano da base de dados}
\usage{
comercioasia_get_ulimoano()
}
\value{
o último ano disponível na base de dados "comerciobr"
}
\description{
Último ano da base de dados
}
