% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_grafico_divisao.R
\name{comercioasia_grafico_divisao}
\alias{comercioasia_grafico_divisao}
\title{Gráfico de Fluxo de Comércio Brasil-país}
\usage{
comercioasia_grafico_divisao(periodo)
}
\arguments{
\item{periodo}{"anual" ou "mensal"}

\item{pais}{um país}
}
\description{
Com base na função \code{comerciobr_dados_corrente}.
}
