% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_grafico_produtos_ranking.R
\name{comercioasia_grafico_produtos_ranking}
\alias{comercioasia_grafico_produtos_ranking}
\title{Gráfico de evolução dos produtos comercializados do Brasil com um determinado país}
\usage{
comercioasia_grafico_produtos_ranking(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gráfico de evolução dos produtos comercializados do Brasil com um determinado país
}
