% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_grafico_valor.R
\name{comercioasia_grafico_valor}
\alias{comercioasia_grafico_valor}
\title{Grafico dos dados extraidos da base_omc}
\usage{
comercioasia_grafico_valor(pais1, pais2)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual"}
}
\description{
Grafico dos dados extraidos da base_omc
}
