% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_grafico_variacao.R
\name{comercioasia_grafico_variacao}
\alias{comercioasia_grafico_variacao}
\title{Dados extraidos da base_omc}
\usage{
comercioasia_grafico_variacao(pais1, pais2)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual"}
}
\description{
Dados extraidos da base_omc
}
