% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_tabela_corrente.R
\name{comercioasia_tabela_corrente}
\alias{comercioasia_tabela_corrente}
\title{Tabela de Corrente de Comércio}
\usage{
comercioasia_tabela_corrente(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
em formato latex.
}
