% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_tabela_divisao.R
\name{comercioasia_tabela_divisao}
\alias{comercioasia_tabela_divisao}
\title{Tabela de Corrente de Comércio}
\usage{
comercioasia_tabela_divisao(periodo)
}
\arguments{
\item{periodo}{"anual" ou "mensal"}

\item{pais}{um país}
}
\description{
em formato latex.
}
