% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_tabela_produtos.R
\name{comercioasia_tabela_produtos}
\alias{comercioasia_tabela_produtos}
\title{Tabela de fluxo comercial por produtos}
\usage{
comercioasia_tabela_produtos(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Tabela de fluxo comercial por produtos
}
