% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comercioasia_tabela_valores.R
\name{comercioasia_tabela_valores}
\alias{comercioasia_tabela_valores}
\title{Tabela dos dados extraidos da base_omc}
\usage{
comercioasia_tabela_valores(pais1, pais2)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual"}
}
\description{
Tabela dos dados extraidos da base_omc
}
